fix: fuel provider initialization
