+++ src/global-constants.ts
@@ -1,3 +1,10 @@
+import { defaultConnectors } from '@fuels/connectors';
+import { coinbaseWallet, injected, walletConnect } from '@wagmi/connectors';
+import { CHAIN_IDS, type FuelConfig, Provider } from 'fuels';
+import { http, createConfig } from "@wagmi/core";
+import { mainnet, sepolia } from "@wagmi/core/chains";
+import { QueryClient } from '@tanstack/react-query';
+
 export const THUNDER_THEME_NAME = "thunder_theme";
 export const WALLET_CONNECT_PROJECT_ID = "fbbe076e89456ef4f6f54493682058b9";
 
@@ -176,3 +183,60 @@ export const wethABI = [
 ];
 
 export const FINALLY_MAINNET = "finally_mainnet";
+
+const FUEL_URLS = {
+  [CHAIN_IDS.fuel.testnet]: "https://testnet.fuel.network/v1/graphql",
+};
+
+// Should be consistent accross project
+export const FUEL_PROVIDER_URL = FUEL_URLS[CHAIN_IDS.fuel.testnet];
+
+export const FUEL_NETWORKS = [
+  {
+    chainId: CHAIN_IDS.fuel.testnet,
+    url: FUEL_PROVIDER_URL,
+  },
+];
+const METADATA = {
+  name: "Wallet Demo",
+  description: "Fuel Wallets Demo",
+  url: location.href,
+  icons: ["https://connectors.fuel.network/logo_white.png"],
+};
+export const WAGMI_CONFIG = createConfig({
+  chains: [mainnet, sepolia],
+  transports: {
+    [mainnet.id]: http(),
+    [sepolia.id]: http(),
+  },
+  connectors: [
+    injected({ shimDisconnect: false }),
+    walletConnect({
+      projectId: WALLET_CONNECT_PROJECT_ID,
+      metadata: METADATA,
+      showQrModal: false,
+    }),
+    coinbaseWallet({
+      appName: METADATA.name,
+      appLogoUrl: METADATA.icons[0],
+      darkMode: true,
+      reloadOnDisconnect: true,
+    }),
+  ],
+});
+export const FUEL_PROVIDER = Provider.create(FUEL_PROVIDER_URL);
+export const FUEL_CONFIG: FuelConfig = {
+
+    connectors: defaultConnectors({
+      // devMode: true,
+      wcProjectId: WALLET_CONNECT_PROJECT_ID,
+      ethWagmiConfig: WAGMI_CONFIG,
+      devMode: true,
+      chainId: CHAIN_IDS.fuel.testnet,
+      fuelProvider: FUEL_PROVIDER,
+    }),
+
+
+}
+
+export const QUERY_CLIENT = new QueryClient();
