+++ src/index.tsx
@@ -11,38 +11,28 @@ import i18next from "i18next";
 import { initReactI18next } from "react-i18next";
 import LOCALES from "./locales";
 
-import { coinbaseWallet, walletConnect } from "@wagmi/connectors";
-import { http, createConfig, injected } from "@wagmi/core";
-import { mainnet, sepolia } from "@wagmi/core/chains";
 
 import * as Sentry from "@sentry/react";
-import { Fuel, FueletWalletConnector, FuelWalletConnector } from "@fuel-wallet/sdk";
+import { Fuel } from 'fuels';
 import { FuelProvider } from "@fuels/react";
 
-// import { Fuel } from "fuels";
-// import { FuelWalletConnector, FueletWalletConnector } from "@fuels/connectors";
+import { FuelWalletConnector, FueletWalletConnector } from "@fuels/connectors";
+import { BrowserTracing } from "@sentry/browser";
 
 export const FuelInstance = new Fuel({
   connectors: [new FueletWalletConnector(), new FuelWalletConnector()],
 });
 
-export const getFuel = () => {
-  return FuelInstance;
-};
-export const getFuelet = () => {
-  return FuelInstance;
-};
-import { BrowserTracing } from "@sentry/browser";
 
 // import { CoinbaseWalletConnector } from "wagmi/connectors/coinbaseWallet";
 import { createClient, reservoirChains } from "@reservoir0x/reservoir-sdk";
 import { LitNodeClient } from "@lit-protocol/lit-node-client";
 import { StytchProvider } from "@stytch/react";
 import { StytchUIClient } from "@stytch/vanilla-js";
-import { WALLET_CONNECT_PROJECT_ID } from "global-constants";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { defaultConnectors } from "@fuels/connectors";
+import { FUEL_NETWORKS, FUEL_CONFIG, WAGMI_CONFIG, QUERY_CLIENT } from "global-constants";
+import {  QueryClientProvider } from "@tanstack/react-query";
 import ModalLogin from "./modal-login";
+import { WagmiProvider } from 'wagmi';
 
 const isDevelopment = "development" === process.env.NODE_ENV;
 
@@ -60,7 +50,7 @@ createClient({
   ],
   apiKey: "151eda75-a312-5fcd-bfac-cf3ea796cb28",
   source: "thundernft.market",
-  marketplaceFees: [`0x313A0A2338999cf88F3F7FF935fe9D9128FFB074:200`],
+  marketplaceFees: ['0x313A0A2338999cf88F3F7FF935fe9D9128FFB074:200'],
 });
 
 i18next.use(initReactI18next).init({
@@ -88,57 +78,23 @@ if (!isDevelopment) {
 
 const stytch = new StytchUIClient("public-token-test-af22c4d3-1e8a-4fa5-a0fa-4c032e2a840a");
 
-const WC_PROJECT_ID = WALLET_CONNECT_PROJECT_ID;
-const METADATA = {
-  name: "Wallet Demo",
-  description: "Fuel Wallets Demo",
-  url: location.href,
-  icons: ["https://connectors.fuel.network/logo_white.png"],
-};
-const wagmiConfig = createConfig({
-  chains: [mainnet, sepolia],
-  transports: {
-    [mainnet.id]: http(),
-    [sepolia.id]: http(),
-  },
-  connectors: [
-    injected({ shimDisconnect: false }),
-    walletConnect({
-      projectId: WC_PROJECT_ID,
-      metadata: METADATA,
-      showQrModal: false,
-    }),
-    coinbaseWallet({
-      appName: METADATA.name,
-      appLogoUrl: METADATA.icons[0],
-      darkMode: true,
-      reloadOnDisconnect: true,
-    }),
-  ],
-});
-const queryClient = new QueryClient();
 
 // eslint-disable-next-line react/no-deprecated
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
       <StytchProvider stytch={stytch}>
-        {/* <WagmiProvider config={WagmiConfig}> */}
-        <QueryClientProvider client={queryClient}>
+        <WagmiProvider config={WAGMI_CONFIG}>
+        <QueryClientProvider client={QUERY_CLIENT}>
           <FuelProvider
             theme="dark"
-            fuelConfig={{
-              connectors: defaultConnectors({
-                // devMode: true,
-                wcProjectId: WALLET_CONNECT_PROJECT_ID,
-                ethWagmiConfig: wagmiConfig,
-              }),
-            }}
+            fuelConfig={FUEL_CONFIG}
+            networks={FUEL_NETWORKS}
           >
             <Router />
           </FuelProvider>
         </QueryClientProvider>
-        {/* </WagmiProvider> */}
+        </WagmiProvider>
       </StytchProvider>
     </Provider>
     <ToastContainer />
